house/requirements.txt
paho-mqtt
requests
Faker
pyshp
shapely
print-color
aiofiles